+++ hardpy/__init__.py
@@ -2,8 +2,12 @@
 # GNU General Public License v3.0 (see LICENSE or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 from hardpy.pytest_hardpy.result import CouchdbLoader
-from hardpy.pytest_hardpy.utils import DuplicateSerialNumberError
+from hardpy.pytest_hardpy.utils import (
+    DuplicateSerialNumberError,
+    DuplicateDialogBoxError,
+)
 from hardpy.pytest_hardpy.result.couchdb_config import CouchdbConfig
+from hardpy.pytest_hardpy.utils import dialog_box
 from hardpy.pytest_hardpy.pytest_call import (
     get_current_report,
     set_dut_info,
@@ -14,12 +18,14 @@ from hardpy.pytest_hardpy.pytest_call import (
     set_run_artifact,
     set_message,
     set_driver_info,
+    run_dialog_box,
 )
 
 __all__ = [
     "CouchdbLoader",
     "CouchdbConfig",
     "DuplicateSerialNumberError",
+    "DuplicateDialogBoxError",
     "get_current_report",
     "set_dut_info",
     "set_dut_serial_number",
@@ -29,4 +35,6 @@ __all__ = [
     "set_run_artifact",
     "set_message",
     "set_driver_info",
+    "run_dialog_box",
+    "dialog_box",
 ]